Make sure the storage types (weak, strong, etc...) are correct. Adding some more generics metadata. Add some nullability metadata.

+++ Source/DKSweptAngleGradient.h
@@ -7,7 +7,7 @@
 #import "DKGradient.h"
 
 typedef union {
-	unsigned long pixel;
+	unsigned int pixel;
 	struct
 		{
 		unsigned char a;
